InvariantPointAttention

Documentation for InvariantPointAttention.

InvariantPointAttention.BackboneUpdate

InvariantPointAttention.IPA

InvariantPointAttention.IPACache

InvariantPointAttention.IPAStructureModuleLayer

InvariantPointAttention.IPCrossA

InvariantPointAttention.IPCrossAStructureModuleLayer

InvariantPointAttention.IPA_settings

InvariantPointAttention.get_T

InvariantPointAttention.get_T_batch

InvariantPointAttention.get_rotation

InvariantPointAttention.get_translation

InvariantPointAttention.left_to_right_mask

InvariantPointAttention.right_to_left_mask

InvariantPointAttention.right_to_left_mask

InvariantPointAttention.softmax1

InvariantPointAttention.BackboneUpdate — Type

Projects the frame embedding => 6, and uses this to transform the input frames.

InvariantPointAttention.IPA — Type

Strictly Self-IPA initialization

InvariantPointAttention.IPACache — Method

IPACache(settings, batchsize)

Initialize an empty IPA cache.

InvariantPointAttention.IPAStructureModuleLayer — Type

Self IPA Partial Structure Module initialization - single layer - adapted from AF2.

InvariantPointAttention.IPCrossA — Type

IPCrossA(settings)

Invariant Point Cross Attention (IPCrossA). Information flows from L (Keys, Values) to R (Queries).

Get settings with IPA_settings

InvariantPointAttention.IPCrossAStructureModuleLayer — Type

Cross IPA Partial Structure Module initialization - single layer - adapted from AF2. From left to right.

InvariantPointAttention.IPA_settings — Method

IPA_settings(
    dims;
    c = 16,
    N_head = 12,
    N_query_points = 4,
    N_point_values = 8,
    c_z = 0,
    Typ = Float32,
    use_softmax1 = false,
    scaling_qk = :default,
)

Returns a tuple of the IPA settings, with defaults for everything except dims. This can be passed to the IPA and IPCrossAStructureModuleLayer.

InvariantPointAttention.get_T — Method

get_T(coords::Array{<:Real, 3})

Get the assosciated SE(3) frame for all residues in a protein backbone represented as a 3x3xL array of coordinates.

InvariantPointAttention.get_T_batch — Method

Get the associated SE(3) frames for all residues in a batch of proteins

InvariantPointAttention.get_rotation — Method

get_rotation([T=Float32,] dims...)

Generates random rotation matrices of given size.

InvariantPointAttention.get_translation — Method

get_translation([T=Float32,] dims...)

Generates random translations of given size.

InvariantPointAttention.left_to_right_mask — Method

left_to_right_mask([T=Float32,] L::Integer, R::Integer; step::Integer = 10)

InvariantPointAttention.right_to_left_mask — Method

right_to_left_mask([T=Float32,] L::Integer, R::Integer; step::Integer = 10)

InvariantPointAttention.right_to_left_mask — Method

right_to_left_mask([T=Float32,] N::Integer)

Create a right-to-left mask for the self-attention mechanism. The mask is a matrix of size N x N where the diagonal and the lower triangular part are set to zero and the upper triangular part is set to infinity.

InvariantPointAttention.softmax1 — Method

softmax1(x, dims = 1)

Behaves like softmax, but as though there was an additional logit of zero along dims (which is excluded from the output). So the values will sum to a value between zero and 1.

See https://www.evanmiller.org/attention-is-off-by-one.html