InvariantPointAttention

Documentation for InvariantPointAttention.

InvariantPointAttention.BackboneUpdate

InvariantPointAttention.IPA

InvariantPointAttention.IPAStructureModuleLayer

InvariantPointAttention.IPA_settings

InvariantPointAttention.T_R3

InvariantPointAttention.T_R3_inv

InvariantPointAttention.T_T

InvariantPointAttention.bcds2quats

InvariantPointAttention.get_rotation

InvariantPointAttention.get_translation

InvariantPointAttention.rotmatrix_from_quat

InvariantPointAttention.update_frame

InvariantPointAttention.BackboneUpdate — Type

Projects the frame embedding => 6, and uses this to transform the input frames.

InvariantPointAttention.IPA — Type

Invariant Point Attention

InvariantPointAttention.IPAStructureModuleLayer — Type

Partial Structure Module - single layer - from AF2. Not a faithful repro, and doesn't include the losses etc.

InvariantPointAttention.IPA_settings — Method

Returns a tuple of the IPA settings, with defaults for everything except dims. This can be passed to the IPA and IPAStructureModuleLayer.

InvariantPointAttention.T_R3 — Method

Applies the SE3 transformations T = (rot,trans) ∈ SE(E3)^N to N batches of m points in R3, i.e., mat ∈ R^(3 x m x N) ↦ T(mat) ∈ R^(3 x m x N). Note here that rotations here are represented in matrix form.

InvariantPointAttention.T_R3_inv — Method

Applys the group inverse of the SE3 transformations T = (rot,trans) ∈ SE(3)^N to N batches of m points in R3, i.e., mat ∈ R^(3 x m x N) ↦ T^(-1)(mat) ∈ R^(3 x m x N) such that T(T^-1(mat)) = mat = T^-1(T(mat)). Note here that rotations here are represented in matrix form.

InvariantPointAttention.T_T — Method

Returns the composition of two SE(3) transformations T1 and T2. Note that if T1 = (R1,t1), and T2 = (R2,t2) then T1T2 = (R1R2, R1*t2 + t1). T here is stored as a tuple (R,t).

InvariantPointAttention.bcds2quats — Method

Creates a quaternion (as a vector) from a triplet of values (pirated from Diffusions.jl)

InvariantPointAttention.get_rotation — Method

Gets N random rotation matrices formatted as an array of size 3x3xN.

InvariantPointAttention.get_translation — Method

Gets N random translations formatted as an array of size 3x1xN (for purposes of broadcasting to arrays of size 3 x m x N)

InvariantPointAttention.rotmatrix_from_quat — Method

Returns the rotation matrix form of N flat quaternions.

InvariantPointAttention.update_frame — Method

Takes a 6-dim vec and maps to a rotation matrix and translation vector, which is then applied to the input frames.